Feat: (Positioning) - Add more prominent/always visible image-upload button to the sidebar. - Add automatic scrolling to end of image gallery upon adding a custom image. - Remove old custom image upload button & styling.
