+++ src/components/Gallery/ImageCarousel.scss
@@ -13,8 +13,7 @@
   }
 }
 
-.selector-preview,
-#add-image-button {
+.selector-preview {
   height: 90px;
   min-height: 80px;
   width: 150px;
